Delete files: a.bat b.bat c.bat
These files is to build C and ASM source code, but they are under the project folder.
Later I found these *.bat may not be with C and ASM files, and also could build source code.
So I delete them, and create them outside the project folder.
And after building, the *.exe and *.obj files also dont have to be with the source code, but anywhere you can select.